Implement audit recommendations
