+++ src/ui/components/audio_control_bar.py
